+++ web/src/App.vue
@@ -9,8 +9,7 @@
   </suspense>
 </template>
 
-<style scoped>
-
+<style lang="scss" scoped>
 .logo {
   height: 6em;
   padding: 1.5em;
