+++ web/src/assets/style/index.scss
@@ -2,15 +2,16 @@
   width: 100%;
   height: 100%;
 }
+
 .el-menu-vertical-demo:not(.el-menu--collapse) {
-  width: 200px;
-  min-height: 400px;
+  width: 194px;
 }
 
 .el-menu-item {
-  border-radius: 30px;
+  border-radius: 32px;
 }
-.el-menu-item.is-active{
+
+.el-menu-item.is-active {
   background: var(--div-bgColor);
 }
 
@@ -59,6 +60,7 @@
   background: transparent;
   border-radius: 25px;
 }
+
 // .el-table .el-table__inner-wrapper {
 //   // height: 100% !important;
 // }
@@ -72,12 +74,14 @@
   // backdrop-filter: saturate(180%) blur(5px);
 }
 
-.el-table .el-table_1_column_1 {
-  padding: 0 0 0 15px;
+.el-table tr th.el-table__cell:first-child,
+.el-table .el-table__body tr td.el-table__cell:first-child {
+  text-align: center;
 }
 
 // 面包屑
-.el-breadcrumb__inner a, .el-breadcrumb__inner.is-link {
+.el-breadcrumb__inner a,
+.el-breadcrumb__inner.is-link {
   color: var(--el-breadcrumb-color)
 }
 
@@ -88,10 +92,61 @@
   color: var(--el-breadcrumb-last-color)
 }
 
+.el-input .el-input__wrapper {
+  border-radius: 18px;
+  background-color: transparent;
+  background-color: var(--div-bgColor);
+  border: 3px solid var(--theme-left-color);
+  box-shadow: none;
+  &.is-focus {
+    border: 3px solid var(--theme-left-color);
+    box-shadow: 0 0 13px 3px var(--theme-left-color);
+  }
+}
+
+
+// 滚动条美化
+* {
+
+  /*滚动条宽或高*/
+  &::-webkit-scrollbar {
+    width: 6px;
+    height: 6px;
+  }
+
+  /*正常情况下滑块的样式*/
+  &::-webkit-scrollbar-thumb {
+    
+  }
+
+  /*鼠标悬浮在该类指向的控件上时滑块的样式*/
+  &:hover::-webkit-scrollbar-thumb {
+    background-color: #9093995e;
+    border-radius: 3px;
+  }
+
+  /*鼠标悬浮在滑块上时滑块的样式*/
+  &::-webkit-scrollbar-thumb:hover {
+    background-color: #90939983;
+    border-radius: 3px;
+  }
+
+  /*正常时候的主干部分*/
+  &::-webkit-scrollbar-track {
+    border-radius: 5px;
+    background-color: transparent;
+  }
+
+  /*鼠标悬浮在滚动条上的主干部分*/
+  &::-webkit-scrollbar-track:hover {
+    background-color: transparent;
+  }
+}
+
 @media (prefers-color-scheme: dark) {
   :root {
-    --el-aside-bgColor : #545c64;
-    --el-main-bgColor  : #f5f5f5;
+    --el-aside-bgColor: #545c64;
+    --el-main-bgColor: #f5f5f5;
     --el-footer-bgColor: #2e3338;
   }
 }
@@ -99,14 +154,17 @@
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
-    --el-aside-bgColor : transparent;
-    --el-main-bgColor  : transparent;
+    --theme-left-color: #b06ab3;
+    --theme-right-color: #4568dc;
+    --el-aside-bgColor: transparent;
+    --el-main-bgColor: transparent;
     --el-footer-bgColor: transparent;
-    --el-card-bgColor  : hsla(0, 0%, 100%, .8);
-    --el-menu-bg-color : transparent;
+    --el-card-bgColor: hsla(0, 0%, 100%, .8);
+    --el-menu-bg-color: transparent;
     --el-breadcrumb-last-color: #ebebeb;
     --el-breadcrumb-color: #fff;
     --div-bgColor: hsla(0, 0%, 100%, 0.5);
+    --el-table_column_1-padding: 0 0 0 15px;
   }
 
   a:hover {
@@ -117,11 +175,3 @@
     background-color: #f9f9f9;
   }
 }
-
-// @media (prefers-color-scheme: light) {
-//   :root {
-//     color: #213547;
-//     background-color: #ffffff;
-//   }
-
-// }
