+++ web/src/main.js
@@ -1,10 +1,10 @@
 import { createApp } from 'vue'
-import './style.css'
 import App from './App.vue'
 import router from './router'
 import { createPinia } from 'pinia'
 import ElementPlus from 'element-plus'
 import 'element-plus/dist/index.css'
+import './style.css'
 import './assets/style/index.scss'
 import '@/utils/theme'
